requirements.txt
streamlit>=1.31
numpy>=1.24
pandas>=2.0
scikit-learn>=1.2
mendeleev>=0.14
rdkit-pypi>=2022.9.5
torch>=2.1
torch-geometric>=2.4
torch-scatter>=2.1
